Rework testing of trimming of operations with no side effects
The previous test was bogus. The only call it trimmed was glxSwapBuffers
which is actually invalid, (this call definitely does have side effects,
but can easily be trimmed if desired with --trim-spec=drawing).

The new testing here uses a trace from a new program, (source
included), called few-side-effects which simply makes various calls
that do not have side effects, (such as glGetIntegerv and
glIsEnabled).

Then, the testing exercises the two independent mechanisms available
for trimming calls without side effects:

    trim --prune

	Trims calls within the user-specified range without side
	effects.

    trim --deps --trim-spec=no-side-effects

	Trims calls outside the user-specific range without side
	effects.

Three new tests are added here, one for each of the above cases and
one for both combined, (prune, trim, and prune-and-trim).