apitrace-tests: Test multiple "--calls" and/or "--frames" command-line options
Tests to go along with patch:

[PATCH] apitrace: Allow multiple --calls and/or --frames input options

These tests were taken from existing tests with "multi-" prefixed to the name.
The apitrace call is taken from the original with the --call or --frame option
split into multiple options but covering the same ranges. Hence the results
should be the same as the original.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>