apitrace-tests: Test comma-separated CALLSETs in "--calls" and "--frames" command-line options
Tests to go along with patch:

[PATCH] apitrace: Allow comma separated lists in --calls and --frames options

These tests were taken from existing tests with the prefix "mix-"
prepended to the name. The apitrace call in the script is taken
from the original script except the --call or --frame option is separated
by commas into multiple options but covering the same ranges. Hence the
results should be the same as the original test.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>