apitrace-tests: Test comma-separated CALLSETs
Tests to go along with patch:

    apitrace: Parse comma-separated list in commandline

These tests were taken from existing tests and the prefix "mix-"
prepended to the name. The apitrace call in the script is taken
from the original script except the --call or --frame option is
separated by commas instead of multiple options but covering the
same ranges. Hence the results should be the same as the original
test.

Signed-off-by: Lawrence L Love <lawrencex.l.love@intel.com>